Add these files back to be tracked (see below)
When I opened the files from the working area I was met with a notice that said that I needed to open those files in a sketch folder and so it moved the files and then opened them. The files were moved to a folder that was auto created in the folder that contains the .git folder so that answered my question if I could separate the different files into folders and still show the .git or if I had to put a .git in each of the sub-folders.