Delete previous additions to file
I checked out a tag and then it changed the head somehow because I couldn't commit changes to this file without it giving me some weird warnings. I deleted (the last commit I think) the commit of the tag from under the Branch tab but it left some "residue" (parent and head comments that were generated), then I was given a choice to refresh the file from its window because an "external program" modified it somehow, so I chose not to refresh it and potentially lose (or add to) what I had already.  I then saved the file and it showed as being modified so to not having that "residue" information. Also, when I had originally did the checkout, I did not "create a local branch" instead I "opened a read only" something option.